Minor changes to avoid compiler warnings
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15534 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/adscSrc/adsc.cpp
@@ -277,7 +277,7 @@ protected:
     epicsEventId stopTriggerEventId;
     epicsEventId lastImageEventId;
 };
-#define NUM_ADSC_PARAMS (&LAST_ADSC_PARAM - &FIRST_ADSC_PARAM + 1)
+#define NUM_ADSC_PARAMS ((int)(&LAST_ADSC_PARAM - &FIRST_ADSC_PARAM + 1))
 
 
 /** Called when asyn clients call pasynInt32->write().
